Add files via upload
Alice has some cards with numbers written on them. She arranges the cards in increasing order, and lays them out face down in a sequence on a table. She challenges Bob to pick out the card containing a given number by turning over as few cards as possible. 

The inputs are the deck of cards in increasing order and a query number. The output is the lowest index of the query number.